#296: ThreadPoolScore's definition move to env.h
