+++ src/leveldb/db/db_impl.cc
@@ -41,7 +41,6 @@
 #include "util/coding.h"
 #include "util/logging.h"
 #include "util/mutexlock.h"
-#include "util/thread_pool.h"
 
 namespace leveldb {
